Correct Matlab path in bash file to find .m files
When starting Matlab with -nojvm, the startup folder is the one from which it was started. However, when starting with the java virtual machine, the startup folder depends on the user preferences. In order to always work in the same folder, a 'cd folder' is done in Matlab before launching the script, folder being the path to the current working directory returned by 'pwd' in the bash file.